chore: shares wrapper cleanup
