+++ packages/sdk/src/Tools.ts
@@ -1,4 +1,2 @@
 export * as GatedRedemptionQueueSharesWrapper from "./Tools/GatedRedemptionQueueSharesWrapper.js";
 export * as SharesSplitter from "./Tools/SharesSplitter.js";
-export * as SharesWrapper from "./Tools/SharesWrapper.js";
-export * as SharesWrapperFactory from "./Tools/SharesWrapperFactory.js";
